[libra_vm] Refactor libra_vm implementation
