+++ language/libra-vm/src/lib.rs
@@ -102,6 +102,7 @@
 
 #[macro_use]
 extern crate mirai_annotations;
+mod access_path_cache;
 #[macro_use]
 mod counters;
 pub mod data_cache;
@@ -115,9 +116,13 @@ pub mod transaction_metadata;
 #[cfg(test)]
 mod unit_tests;
 
+pub mod libra_transaction_executor;
+pub mod libra_transaction_validator;
 pub mod system_module_names;
 
-pub use crate::libra_vm::{txn_effects_to_writeset_and_events, LibraVM};
+pub use crate::libra_transaction_executor::LibraVM;
+pub use crate::libra_transaction_validator::LibraVMValidator;
+pub use crate::libra_vm::txn_effects_to_writeset_and_events;
 
 use libra_state_view::StateView;
 use libra_types::{
