+++ language/libra-vm/src/unit_tests/block_chunking_test.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::libra_vm::{chunk_block_transactions, TransactionBlock};
+use crate::libra_transaction_executor::{chunk_block_transactions, TransactionBlock};
 use libra_types::transaction::Transaction;
 use proptest::{collection::vec, prelude::*};
