+++ vm-validator/src/vm_validator.rs
@@ -9,7 +9,7 @@ use libra_types::{
     on_chain_config::{LibraVersion, OnChainConfigPayload, VMConfig},
     transaction::{SignedTransaction, VMValidatorResult},
 };
-use libra_vm::LibraVM;
+use libra_vm::LibraVMValidator;
 use scratchpad::SparseMerkleTree;
 use std::{convert::TryFrom, sync::Arc};
 use storage_interface::{state_view::VerifiedStateView, DbReader};
@@ -31,12 +31,12 @@ pub trait TransactionValidation: Send + Sync + Clone {
 #[derive(Clone)]
 pub struct VMValidator {
     db_reader: Arc<dyn DbReader>,
-    vm: LibraVM,
+    vm: LibraVMValidator,
 }
 
 impl VMValidator {
     pub fn new(db_reader: Arc<dyn DbReader>) -> Self {
-        let mut vm = LibraVM::new();
+        let mut vm = LibraVMValidator::new();
         let (version, state_root) = db_reader.get_latest_state_root().expect("Should not fail.");
         let smt = SparseMerkleTree::new(state_root);
         let state_view = VerifiedStateView::new(
@@ -53,7 +53,7 @@ impl VMValidator {
 }
 
 impl TransactionValidation for VMValidator {
-    type ValidationInstance = LibraVM;
+    type ValidationInstance = LibraVMValidator;
 
     fn validate_transaction(&self, txn: SignedTransaction) -> Result<VMValidatorResult> {
         use libra_vm::VMValidator;
@@ -80,7 +80,7 @@ impl TransactionValidation for VMValidator {
         let vm_config = config.get::<VMConfig>()?;
         let version = config.get::<LibraVersion>()?;
 
-        self.vm = LibraVM::init_with_config(version, vm_config);
+        self.vm = LibraVMValidator::init_with_config(version, vm_config);
         Ok(())
     }
 }
